+++ renku/core/management/git.py
@@ -26,6 +26,7 @@ import time
 import uuid
 from contextlib import contextmanager
 from pathlib import Path
+from typing import List
 
 import attr
 
@@ -150,16 +151,15 @@ def prepare_worktree(
     # TODO sys.argv
 
     if commit is NULL_TREE:
-        args = ["add", "--detach", path]
-        original_client.repository.run_git_command("worktree", *args)
+        original_client.repository.create_worktree(path, detach=True)
         client = attr.evolve(original_client, path=path)
         client.repository.run_git_command("checkout", "--orphan", branch_name)
         client.repository.remove("*", recursive=True, force=True)
     else:
-        args = ["add", "-b", branch_name, path]
+        revision = None
         if commit:
-            args.append(commit.hexsha)
-        original_client.repository.run_git_command("worktree", *args)
+            revision = commit.hexsha
+        original_client.repository.create_worktree(path, branch=branch_name, reference=revision)
         client = attr.evolve(original_client, path=path)
 
     client.repository.get_configuration = original_client.repository.get_configuration
@@ -207,7 +207,7 @@ def finalize_worktree(
         raise errors.FailedMerge(client.repository, branch_name, merge_args)
 
     if delete:
-        client.repository.run_git_command("worktree", "remove", path)
+        client.repository.remove_worktree(path)
 
         if new_branch:
             # delete the created temporary branch
@@ -257,9 +257,8 @@ def _clean_streams(repository, mapped_streams):
         if path not in repository.files:
             os.remove(absolute_path)
         else:
-            with open(path, "wb") as output_file:
-                checksum = repository.get_object_hash(path=absolute_path, revision="HEAD")
-                repository.copy_content_to_file(path=absolute_path, checksum=checksum, output_file=output_file)
+            checksum = repository.get_object_hash(path=absolute_path, revision="HEAD")
+            repository.copy_content_to_file(path=absolute_path, checksum=checksum, output_path=path)
 
 
 def _expand_directories(paths):
@@ -321,7 +320,7 @@ class GitCore:
             for path in itertools.chain(self.repository.files, self.repository.untracked_files)
         ]
 
-    def find_ignored_paths(self, *paths):
+    def find_ignored_paths(self, *paths) -> List[str]:
         """Return ignored paths matching ``.gitignore`` file."""
         return self.repository.get_ignored_paths(*paths)