+++ tests/core/fixtures/core_workflow.py
@@ -17,12 +17,16 @@
 # limitations under the License.
 """Renku core fixtures for workflow testing."""
 
+from datetime import datetime, timedelta
+
 import pytest
 
+from renku.domain_model.provenance.activity import Activity
 from renku.domain_model.workflow.composite_plan import CompositePlan
 from renku.domain_model.workflow.parameter import CommandInput, CommandOutput, CommandParameter
 from renku.domain_model.workflow.plan import Plan
-from tests.utils import write_and_commit_file
+from renku.infrastructure.gateway.activity_gateway import ActivityGateway
+from tests.utils import create_dummy_plan
 
 
 def _create_run(name: str) -> Plan:
@@ -31,26 +35,26 @@ def _create_run(name: str) -> Plan:
     input1 = CommandInput(
         id=CommandInput.generate_id(run_id, 1),
         position=1,
-        default_value=1,
+        default_value="1",
         name=f"{name}_input1",
     )
     input2 = CommandInput(
         id=CommandInput.generate_id(run_id, 2),
         position=2,
-        default_value=2,
+        default_value="2",
         name=f"{name}_input2",
     )
 
     output1 = CommandOutput(
         id=CommandOutput.generate_id(run_id, 3),
         position=3,
-        default_value=3,
+        default_value="3",
         name=f"{name}_output1",
     )
     output2 = CommandOutput(
         id=CommandOutput.generate_id(run_id, 4),
         position=4,
-        default_value=4,
+        default_value="4",
         name=f"{name}_output2",
     )
 
@@ -89,11 +93,56 @@ def composite_plan():
 
 
 @pytest.fixture
-def client_with_runs(client, run, client_database_injection_manager):
+def client_with_runs(client, client_database_injection_manager):
     """A client with runs."""
-    write_and_commit_file(client.repository, "input", "some\ninput\nfile")
 
-    assert 0 == run(["run", "--name", "plan-1", "head", "-n", "2", "input"], stdout="intermediate")
-    assert 0 == run(["run", "--name", "plan-2", "tail", "-n", "1", "intermediate"], stdout="output")
+    def create_activity(plan, date, index) -> Activity:
+        """Create an activity with id /activities/index."""
+        return Activity.from_plan(
+            plan=plan,
+            id=Activity.generate_id(str(index)),
+            started_at_time=date,
+            ended_at_time=date + timedelta(seconds=1),
+        )
+
+    date_1 = datetime(2022, 5, 20, 0, 42, 0)
+    date_2 = datetime(2022, 5, 20, 0, 43, 0)
+
+    plan_1 = create_dummy_plan(
+        command="command-1",
+        date_created=date_1,
+        description="First plan",
+        index=1,
+        inputs=["input"],
+        keywords=["plans", "1"],
+        name="plan-1",
+        outputs=[("intermediate", "stdout")],
+        parameters=[("parameter-1", "42", "-n ")],
+        success_codes=[0, 1],
+    )
+
+    plan_2 = create_dummy_plan(
+        command="command-2",
+        date_created=date_2,
+        description="Second plan",
+        index=2,
+        inputs=["intermediate"],
+        keywords=["plans", "2"],
+        name="plan-2",
+        outputs=[("output", "stdout")],
+        parameters=[("int-parameter", 43, "-n "), ("str-parameter", "some value", None)],
+    )
+
+    with client_database_injection_manager(client):
+        activity_1 = create_activity(plan_1, date_1, index=1)
+        activity_2 = create_activity(plan_2, date_2, index=2)
+
+        activity_gateway = ActivityGateway()
+
+        activity_gateway.add(activity_1)
+        activity_gateway.add(activity_2)
+
+    client.repository.add(all=True)
+    client.repository.commit("Add runs")
 
     yield client